@@ -56,114 +57,115 @@
     "universal:watch": "npm run build:universal && npm-run-all -p -r watch:universal server:universal"
   },
   "dependencies": {
-    "@angular/animations": "~5.2.9",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "~5.2.9",
-    "@angular/compiler": "~5.2.9",
-    "@angular/compiler-cli": "~5.2.9",
-    "@angular/core": "~5.2.9",
-    "@angular/forms": "~5.2.9",
-    "@angular/http": "~5.2.9",
-    "@angular/platform-browser": "~5.2.9",
-    "@angular/platform-browser-dynamic": "~5.2.9",
-    "@angular/platform-server": "~5.2.9",
-    "@angular/router": "~5.2.9",
-    "@angular/tsc-wrapped": "~4.4.6",
+    "@angular/animations": "~6.0.1",
+    "@angular/cdk": "^6.0.2",
+    "@angular/common": "~6.0.1",
+    "@angular/compiler": "~6.0.1",
+    "@angular/compiler-cli": "~6.0.1",
+    "@angular/core": "~6.0.1",
+    "@angular/forms": "~6.0.1",
+    "@angular/http": "~6.0.1",
+    "@angular/platform-browser": "~6.0.1",
+    "@angular/platform-browser-dynamic": "~6.0.1",
+    "@angular/platform-server": "~6.0.1",
+    "@angular/router": "~6.0.1",
+    "@angular/tsc-wrapped": "^4.4.7",
     "@angularclass/hmr": "2.1.3",
     "@angularclass/hmr-loader": "3.0.4",
     "@angularclass/idle-preload": "~2.0.1",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0-beta.8",
+    "@ng-bootstrap/ng-bootstrap": "2.0.0",
     "@ngrx/core": "1.2.0",
     "@ngrx/effects": "5.2.0",
     "@ngrx/router-store": "5.2.0",
     "@ngrx/store": "5.2.0",
     "@ngrx/store-devtools": "5.2.0",
     "@ngrx/store-log-monitor": "3.0.2",
-    "@nguniversal/express-engine": "5.0.0-beta.6",
-    "body-parser": "^1.18.2",
-    "bootstrap": "4.0.0-beta.2",
-    "compression": "^1.7.1",
+    "@nguniversal/express-engine": "6.0.0",
+    "body-parser": "^1.18.3",
+    "bootstrap": "4.1.1",
+    "compression": "^1.7.2",
     "cookie-parser": "1.4.3",
-    "core-js": "2.5.3",
-    "express": "~4.16.2",
+    "core-js": "2.5.6",
+    "express": "~4.16.3",
     "ip": "~1.1.5",
     "methods": "1.1.2",
-    "ngrx-store-freeze": "~0.2.1",
-    "ngrx-store-logger": "~0.2.0",
-    "rxjs": "~5.5.7",
+    "ngrx-store-freeze": "~0.2.2",
+    "ngrx-store-logger": "~0.2.1",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
     "script-ext-html-webpack-plugin": "^2.0.1",
     "web-animations-js": "^2.3.1",
-    "zone.js": "~0.8.20"
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@ngtools/webpack": "1.10.2",
-    "@types/body-parser": "~1.16.8",
-    "@types/compression": "~0.0.35",
+    "@ngtools/webpack": "6.0.1",
+    "@types/body-parser": "~1.17.0",
+    "@types/compression": "~0.0.36",
     "@types/cookie-parser": "~1.4.1",
-    "@types/express": "~4.11.0",
-    "@types/express-serve-static-core": "~4.11.0",
-    "@types/ip": "~0.0.30",
-    "@types/jasmine": "2.8.3",
+    "@types/express": "~4.11.1",
+    "@types/express-serve-static-core": "~4.11.1",
+    "@types/ip": "~0.0.31",
+    "@types/jasmine": "2.8.7",
     "@types/mime": "~2.0.0",
     "@types/node": "~8.5.2",
     "@types/selenium-webdriver": "~3.0.8",
-    "@types/serve-static": "~1.13.1",
-    "angular-router-loader": "~0.8.1",
+    "@types/serve-static": "~1.13.2",
+    "angular-router-loader": "~0.8.5",
     "angular2-template-loader": "~0.6.2",
     "apply-loader": "^2.0.0",
-    "awesome-typescript-loader": "~3.1.2",
-    "codelyzer": "~4.0.2",
-    "compression-webpack-plugin": "~1.1.3",
-    "copy-webpack-plugin": "~4.3.1",
-    "css-loader": "~0.28.7",
-    "es6-promise": "~4.2.2",
+    "awesome-typescript-loader": "~5.0.0",
+    "codelyzer": "~4.3.0",
+    "compression-webpack-plugin": "~1.1.11",
+    "copy-webpack-plugin": "~4.5.1",
+    "css-loader": "~0.28.11",
+    "es6-promise": "~4.2.4",
     "es6-promise-loader": "~1.0.2",
     "extract-text-webpack-plugin": "^3.0.2",
-    "html-webpack-plugin": "~2.30.1",
-    "imports-loader": "~0.7.1",
-    "istanbul-instrumenter-loader": "~3.0.0",
+    "html-webpack-plugin": "~3.2.0",
+    "imports-loader": "~0.8.0",
+    "istanbul-instrumenter-loader": "~3.0.1",
     "json-loader": "~0.5.7",
-    "karma": "~2.0.0",
+    "karma": "~2.0.2",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage": "~1.1.1",
-    "karma-jasmine": "~1.1.1",
+    "karma-coverage": "~1.1.2",
+    "karma-jasmine": "~1.1.2",
     "karma-mocha-reporter": "~2.2.5",
     "karma-phantomjs-launcher": "~1.0.4",
-    "karma-remap-coverage": "0.1.4",
+    "karma-remap-coverage": "0.1.5",
     "karma-sourcemap-loader": "~0.3.7",
-    "karma-webpack": "2.0.9",
-    "node-sass": "~4.7.2",
-    "nodemon": "~1.14.7",
-    "npm-run-all": "~4.1.2",
+    "karma-webpack": "3.0.0",
+    "nodemon": "~1.17.4",
+    "npm-run-all": "~4.1.3",
     "phantomjs-polyfill": "~0.0.2",
     "phantomjs-prebuilt": "~2.1.16",
-    "protractor": "~5.2.2",
+    "protractor": "~5.3.2",
     "pug": "^2.0.3",
     "pug-loader": "^2.4.0",
     "raw-loader": "~0.5.1",
     "rimraf": "~2.6.2",
-    "sass-loader": "^6.0.6",
     "source-map-loader": "~0.2.3",
-    "string-replace-loader": "~1.3.0",
+    "string-replace-loader": "~2.1.1",
     "style-loader": "^0.21.0",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.2",
     "to-string-loader": "~1.1.5",
     "ts-helpers": "~1.1.2",
-    "ts-node": "~4.1.0",
-    "tslib": "^1.5.0",
-    "tslint": "~5.8.0",
-    "tslint-loader": "~3.5.3",
-    "typescript": "~2.4.2",
+    "ts-node": "~6.0.3",
+    "tslib": "^1.9.1",
+    "tslint": "~5.10.0",
+    "tslint-loader": "~3.6.0",
+    "typescript": "~2.7.2",
+    "uglifyjs-webpack-plugin": "^1.2.5",
     "watch-run": "~1.2.5",
-    "webpack": "~3.10.0",
-    "webpack-bundle-analyzer": "^2.9.1",
-    "webpack-dev-middleware": "~2.0.3",
-    "webpack-dev-server": "~2.9.7",
+    "webpack": "~4.8.3",
+    "webpack-bundle-analyzer": "^2.11.3",
+    "webpack-cli": "^2.1.3",
+    "webpack-dev-middleware": "~3.1.3",
+    "webpack-dev-server": "^3.1.4",
     "webpack-md5-hash": "^0.0.6",
-    "webpack-merge": "~4.1.1",
-    "webpack-node-externals": "^1.6.0"
+    "webpack-merge": "~4.1.2",
+    "webpack-node-externals": "^1.7.2"
   },
   "license": "MIT",
-  "homepage": "https://js.magora.team/components/ng5"
+  "homepage": "https://js.magora.team/components/ng6"
 }
