@@ -1,7 +1,7 @@
-import 'rxjs/add/observable/of';
+import {of as observableOf,  Observable } from 'rxjs';
+
 import { Injectable } from '@angular/core';
 import { PreloadingStrategy, Route } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
 
 @Injectable()
 export class SelectivePreloadingStrategy implements PreloadingStrategy {
@@ -17,7 +17,7 @@ export class SelectivePreloadingStrategy implements PreloadingStrategy {
 
       return load();
     } else {
-      return Observable.of(null);
+      return observableOf(null);
     }
   }
 }
