@@ -1,10 +1,13 @@
+
+import {from as observableFrom,  Observable } from 'rxjs';
+
+import {tap} from 'rxjs/operators';
 /* tslint:disable: max-line-length no-shadowed-variable */
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
 import { TransferState } from '@angular/platform-browser';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/observable/fromPromise';
+
+
 
 @Injectable()
 export class TransferHttp {
@@ -196,10 +199,10 @@ export class TransferHttp {
     try {
       return this.resolveData(key);
     } catch (e) {
-      return callback(method, uri, options)
-        .do(data => {
+      return callback(method, uri, options).pipe(
+        tap(data => {
           this.setCache(key, data);
-        });
+        }));
     }
   }
 
@@ -219,10 +222,10 @@ export class TransferHttp {
     try {
       return this.resolveData(key);
     } catch (e) {
-      return callback(uri, body, options)
-        .do(data => {
+      return callback(uri, body, options).pipe(
+        tap(data => {
           this.setCache(key, data);
-        });
+        }));
     }
   }
 
@@ -233,7 +236,7 @@ export class TransferHttp {
       throw new Error();
     }
 
-    return Observable.fromPromise(Promise.resolve(data));
+    return observableFrom(Promise.resolve(data));
   }
 
   private setCache(key, data) {
