@@ -1,25 +1,12 @@
 // RxJS
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/combineLatest';
-import 'rxjs/add/observable/throw';
-import 'rxjs/add/operator/debounceTime';
-import 'rxjs/add/operator/distinctUntilChanged';
-import 'rxjs/add/operator/filter';
 import 'rxjs/add/operator/first';
-import 'rxjs/add/observable/from';
-import 'rxjs/add/observable/fromEvent';
-import 'rxjs/add/observable/interval';
-import 'rxjs/add/operator/let';
+import 'rxjs/add/operator/filter';
 import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/mapTo';
-import 'rxjs/add/operator/mergeMap';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/pluck';
-import 'rxjs/add/operator/sample';
 import 'rxjs/add/operator/switchMap';
+import 'rxjs/add/operator/mergeMap';
 import 'rxjs/add/operator/take';
+import 'rxjs/add/operator/catch';
 import 'rxjs/add/operator/takeUntil';
-import 'rxjs/add/operator/toArray';
 import 'rxjs/ReplaySubject';
 import 'rxjs/Observable';
 import 'rxjs/Subject';
